+++ getOperatorTie.c
@@ -1,15 +1,15 @@
 #define _CRT_SECURE_NO_WARNINGS
 #include "calculator.h"
 
-int getOperatorTie(expType operator) {
+byte getOperatorTie(expType operator) {
 
 	//0 - left 1 - right
 
 	switch (operator) {
-	case power:
-		return 1;
-	default:
-		return 0;
+		case power:
+			return 1;
+		default:
+			return 0;
 	}
 
 	return -1;
